+++ tests/parser/test_address_parser_test_api.py
@@ -1,6 +1,7 @@
 # Since we use a patch as model mock we skip the unused argument error
 # pylint: disable=unused-argument, too-many-arguments
 import unittest
+from unittest import skipIf
 from unittest.mock import patch, call
 
 import torch
@@ -15,9 +16,7 @@ class AddressParserRetrainTest(AddressParserPredictTestCase):
     @classmethod
     def setUpClass(cls):
         super(AddressParserRetrainTest, cls).setUpClass()
-        cls.a_device = "cpu"
-
-        cls.a_device = "cpu"
+        cls.a_device = torch.device("cpu")
 
         cls.a_train_ratio = 0.8
         cls.a_batch_size = BATCH_SIZE
@@ -26,7 +25,7 @@ class AddressParserRetrainTest(AddressParserPredictTestCase):
         cls.a_learning_rate = 0.01
         cls.a_callbacks_list = []
         cls.a_seed = 42
-        cls.a_torch_device = torch.device(cls.a_device)
+        cls.a_torch_device = torch.device("cuda:0")
 
         cls.a_loss_function = nll_loss
         cls.a_list_of_batch_metrics = [accuracy]
@@ -45,11 +44,11 @@ class AddressParserRetrainTest(AddressParserPredictTestCase):
                                  callbacks=self.a_callbacks_list,
                                  seed=self.a_seed)
 
-    def assert_experiment_test(self, experiment_mock, model_mock):
+    def assert_experiment_test(self, experiment_mock, model_mock, device):
         experiment_mock.assert_called_with(
             "./checkpoint",  # we always use this as default logging dir.
             model_mock(),
-            device=self.a_torch_device,
+            device=device,
             loss_function=self.a_loss_function,
             batch_metrics=self.a_list_of_batch_metrics,
             logging=False)
@@ -67,7 +66,7 @@ class AddressParserRetrainTest(AddressParserPredictTestCase):
     @patch("deepparse.parser.address_parser.FastTextVectorizer")
     @patch("deepparse.parser.address_parser.FastTextEmbeddingsModel")
     @patch("deepparse.parser.address_parser.download_fasttext_embeddings")
-    def test_givenAFasttextModel_whenTest_thenInstantiateExperimentProperly(
+    def test_givenAFasttextModel_whenTestCPU_thenInstantiateExperimentProperly(
             self, download_weights_mock, embeddings_model_mock, vectorizer_model_mock, data_padding_mock, model_mock,
             data_transform_mock, optimizer_mock, experiment_mock, dataloader_mock):
         self.address_parser = AddressParser(model_type=self.a_fasttext_model_type,
@@ -75,7 +74,27 @@ class AddressParserRetrainTest(AddressParserPredictTestCase):
                                             verbose=self.verbose)
         self.address_parser_test_call()
 
-        self.assert_experiment_test(experiment_mock, model_mock)
+        self.assert_experiment_test(experiment_mock, model_mock, device=self.a_device)
+
+    @patch("deepparse.parser.address_parser.DataLoader")
+    @patch("deepparse.parser.address_parser.Experiment")
+    @patch("deepparse.parser.address_parser.SGD")
+    @patch("deepparse.parser.address_parser.DataTransform")
+    @patch("deepparse.parser.address_parser.FastTextSeq2SeqModel")
+    @patch("deepparse.parser.address_parser.fasttext_data_padding")
+    @patch("deepparse.parser.address_parser.FastTextVectorizer")
+    @patch("deepparse.parser.address_parser.FastTextEmbeddingsModel")
+    @patch("deepparse.parser.address_parser.download_fasttext_embeddings")
+    @skipIf(not torch.cuda.is_available(), "no gpu available")
+    def test_givenAFasttextModel_whenTestGPU_thenInstantiateExperimentProperly(
+            self, download_weights_mock, embeddings_model_mock, vectorizer_model_mock, data_padding_mock, model_mock,
+            data_transform_mock, optimizer_mock, experiment_mock, dataloader_mock):
+        self.address_parser = AddressParser(model_type=self.a_fasttext_model_type,
+                                            device=self.a_torch_device,
+                                            verbose=self.verbose)
+        self.address_parser_test_call()
+
+        self.assert_experiment_test(experiment_mock, model_mock, device=self.a_torch_device)
 
     @patch("deepparse.parser.address_parser.DataLoader")
     @patch("deepparse.parser.address_parser.Experiment")
@@ -124,17 +143,38 @@ class AddressParserRetrainTest(AddressParserPredictTestCase):
     @patch("deepparse.parser.address_parser.bpemb_data_padding")
     @patch("deepparse.parser.address_parser.BPEmbVectorizer")
     @patch("deepparse.parser.address_parser.BPEmbEmbeddingsModel")
-    def test_givenABPEmbModel_whenTest_thenInstantiateExperimentProperly(self, embeddings_model_mock,
-                                                                         vectorizer_model_mock, data_padding_mock,
-                                                                         model_mock, data_transform_mock,
-                                                                         optimizer_mock, experiment_mock,
-                                                                         dataloader_mock):
+    def test_givenABPEmbModel_whenTestCPU_thenInstantiateExperimentProperly(self, embeddings_model_mock,
+                                                                            vectorizer_model_mock, data_padding_mock,
+                                                                            model_mock, data_transform_mock,
+                                                                            optimizer_mock, experiment_mock,
+                                                                            dataloader_mock):
         self.address_parser = AddressParser(model_type=self.a_bpemb_model_type,
                                             device=self.a_device,
                                             verbose=self.verbose)
         self.address_parser_test_call()
 
-        self.assert_experiment_test(experiment_mock, model_mock)
+        self.assert_experiment_test(experiment_mock, model_mock, device=self.a_device)
+
+    @patch("deepparse.parser.address_parser.DataLoader")
+    @patch("deepparse.parser.address_parser.Experiment")
+    @patch("deepparse.parser.address_parser.SGD")
+    @patch("deepparse.parser.address_parser.DataTransform")
+    @patch("deepparse.parser.address_parser.BPEmbSeq2SeqModel")
+    @patch("deepparse.parser.address_parser.bpemb_data_padding")
+    @patch("deepparse.parser.address_parser.BPEmbVectorizer")
+    @patch("deepparse.parser.address_parser.BPEmbEmbeddingsModel")
+    @skipIf(not torch.cuda.is_available(), "no gpu available")
+    def test_givenABPEmbModel_whenTestGPU_thenInstantiateExperimentProperly(self, embeddings_model_mock,
+                                                                            vectorizer_model_mock, data_padding_mock,
+                                                                            model_mock, data_transform_mock,
+                                                                            optimizer_mock, experiment_mock,
+                                                                            dataloader_mock):
+        self.address_parser = AddressParser(model_type=self.a_bpemb_model_type,
+                                            device=self.a_torch_device,
+                                            verbose=self.verbose)
+        self.address_parser_test_call()
+
+        self.assert_experiment_test(experiment_mock, model_mock, device=self.a_torch_device)
 
     @patch("deepparse.parser.address_parser.DataLoader")
     @patch("deepparse.parser.address_parser.Experiment")