+++ CHANGELOG.md
@@ -96,3 +96,4 @@
 
 - Added `__eq__` method to `FormattedParsedAddress`.
 - Improved device management.
+- Improved testing.